feat(variant): SKFP-955 switch url v2 (#3951)
