+++ src/App.tsx
@@ -45,7 +45,6 @@ const Studies = loadable(() => import('views/Studies'), loadableProps);
 const DataExploration = loadable(() => import('views/DataExploration'), loadableProps);
 const Variants = loadable(() => import('views/Variants'), loadableProps);
 const VariantEntity = loadable(() => import('views/VariantEntity'), loadableProps);
-const VariantEntity2 = loadable(() => import('views/VariantEntity2'), loadableProps);
 const FileEntity = loadable(() => import('views/FileEntity'), loadableProps);
 const ProfileSettings = loadable(() => import('views/Profile/Settings'), loadableProps);
 
@@ -205,15 +204,6 @@ const App = () => {
                     element={
                       <ProtectedRoute>
                         <VariantEntity />
-                      </ProtectedRoute>
-                    }
-                  />
-
-                  <Route
-                    path={DYNAMIC_ROUTES.VARIANT_ENTITY_2}
-                    element={
-                      <ProtectedRoute>
-                        <VariantEntity2 />
                       </ProtectedRoute>
                     }
                   />
