--- src/views/VariantEntity/SummaryHeader/index.module.scss
@@ -1,60 +1,0 @@
-@import 'src/style/themes/kids-first/colors';
-@import 'src/style/themes/kids-first/theme.template';
-
-.container {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 20px;
-  gap: 24px;
-}
-
-.link {
-  display: flex;
-  padding: 8px 16px;
-  border: 1px solid $gray-5;
-  width: 100%;
-  align-items: center;
-  justify-content: center;
-
-  &:hover {
-    background: $gray-2;
-  }
-}
-
-.participantDisabled {
-  display: flex;
-  height: 54px;
-  padding: 8px 16px;
-  border: 1px solid $gray-5;
-  width: 100%;
-  align-items: center;
-  justify-content: center;
-
-  &:hover {
-    background-color: $gray-1;
-  }
-}
-
-.icon,
-.entityCount {
-  color: $geekblue-7;
-}
-
-.icon {
-  font-size: 20px;
-  margin-bottom: 3px;
-}
-
-.entityCount {
-  font-size: 24px;
-  line-height: 32px;
-  font-weight: 600;
-  margin: 0 4px 0 8px;
-}
-
-.text {
-  color: $gray-7;
-  line-height: 32px;
-  margin-top: 4px;
-}
