+++ src/views/VariantEntity/utils/consequence.tsx
